phobos 0.153

+++ object.d
@@ -8,7 +8,7 @@ alias bool bit;
 
 alias typeof(int.sizeof) size_t;
 alias typeof(cast(void*)0 - cast(void*)0) ptrdiff_t;
-alias uint hash_t;
+alias size_t hash_t;
 
 extern (C)
 {   int printf(char *, ...);
